Added Tests

+++ ASP.NET_Uni_Project/Controllers/AuctionsController.cs
@@ -1,5 +1,7 @@
 ﻿using ASP.NET_Uni_Project.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
+using Microsoft.EntityFrameworkCore;
 
 namespace ASP.NET_Uni_Project.Controllers
 {
@@ -15,30 +17,53 @@ namespace ASP.NET_Uni_Project.Controllers
 
         //GET api/aucitons
         [HttpGet]
-        public IEnumerable<Auction> Get()
+        public ActionResult<IEnumerable<Auction?>> GetAuctions()
         {
-            return _auctionService.FindAll();
+            if (_auctionService is null)
+            {
+                return NotFound();
+            }
+            return new ActionResult<IEnumerable<Auction?>>(_auctionService.FindAll());
         }
 
         //GET api/auctions/{Id}
-        [HttpGet("{id}", Name = "Get")]
-        public ActionResult<Auction> Get(int id)
+        [HttpGet("{id}")]
+        public ActionResult<Auction?> GetAuction(int id)
         {
+            if (_auctionService is null)
+            {
+                return NotFound();
+            }
             return _auctionService.FindById(id);
         }
 
         //POST api/auctions/
         [HttpPost]
-        public ActionResult Post([FromBody] Auction auction)
+        public ActionResult PostAuction([FromBody] Auction auction)
         {
-            _auctionService.Create(auction);
-            return Created("", auction);
+            if (_auctionService == null)
+            {
+                return Problem("Service Not Found.");
+            }
+            if (ModelState.IsValid)
+            {
+                var createdId = _auctionService.Create(auction);
+                return Created($"/api/auctions/{createdId}", createdId);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         //DELETE api/auctions/{Id}
         [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public IActionResult DeleteAuction(int id)
         {
+            if (_auctionService == null)
+            {
+                return NotFound();
+            }
             var result = _auctionService.Delete(id);
             if (result == null)
             {
@@ -49,14 +74,47 @@ namespace ASP.NET_Uni_Project.Controllers
 
         //PUT api/auctions/{Id}
         [HttpPut("{id}")]
-        public ActionResult Put(int id, [FromBody] Auction auction)
+        public ActionResult PutAuction(int id, [FromBody] Auction auction)
         {
-            auction.Id = (int)id;
-            if (_auctionService.Update(auction))
+            if (id != auction.Id)
             {
                 return BadRequest();
             }
+            auction.Id = id;
+            _auctionService.Update(auction);
             return NoContent();
+        }
+
+        //PUT api/auctions/{Id}/bid
+        [HttpPut]
+        [Route("{id}/bid")]
+        public ActionResult PutBid([FromRoute] int id, [FromBody] Auction auction)
+        {
+            if (_auctionService.FindById(id) is null)
+            {
+                return NotFound();
+            }
+            if (_auctionService.DoBid(auction))
+            {
+                return Ok();
+            }
+            return Problem("Couldn't put bid.");
+        }
+
+        //PUT api/auctions/{Id}/buy
+        [HttpPut]
+        [Route("{id}/buy")]
+        public ActionResult PutBuy([FromRoute] int id, [FromBody] Auction auction)
+        {
+            if (_auctionService.FindById(id) is null)
+            {
+                return NotFound();
+            }
+            if (_auctionService.DoBuy(auction))
+            {
+                return Ok();
+            }
+            return Problem("Couldn't buy the item.");
         }
     }
 }
